pyproject.toml
[tool.poetry]
version = "0.1.0"
description = "stamps e-store"
authors = ["Miron Sadykov <MironSadykov@yandex.ru>"]
license = "MIT"
readme = "docs/README.md"
repository = "https://github.com/Reagent992/stamps"
name = "src"
packages = [{include = "docs"}]


[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.6"
python-slugify = "^8.0.1"
sorl-thumbnail = "^12.10.0"
django-view-breadcrumbs = "^2.4.1"
pillow = "^10.0.1"
django-ckeditor-5 = "^0.2.10"
environs = {extras = ["django"], version = "^9.5.0"}
django-json-widget = "^2.0.1"
django-crispy-forms = "^2.1"
crispy-bootstrap5 = "^2024.2"
celery = "^5.4.0"
django-dirtyfields = "^1.9.2"
flower = "^2.0.1"
psycopg2-binary = "^2.9.9"
djlint = "^1.34.1"

[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.2.0"
black = "^23.9.1"
isort = "^5.12.0"
flake8 = "^6.1.0"
flake8-broken-line = "^1.0.0"
flake8-plugin-utils = "^1.3.3"
flake8-return = "^1.2.0"
pre-commit = "^3.5.0"
ipython = "^8.24.0"
django-stubs = {extras = ["compatible-mypy"], version = "^5.0.0"}
ipykernel = "^6.29.4"

[tool.poetry.scripts]
dr = "docs.poetry_scripts:dr"
ds = "docs.poetry_scripts:ds"
cw = "docs.poetry_scripts:cw"
rb = "docs.poetry_scripts:rb"
flower = "docs.poetry_scripts:flower"
dt = "docs.poetry_scripts:dt"
rc = "docs.poetry_scripts:rc"




[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
(
    \.git
    | venv
    | \.venv
    | _build
    | buck-out
    | build
    | migrations
    | settings
    | wsgi
    | asgi
    | manage
    | \.pyc
    | \.pyo
    | \.pyd
    | \.db
    | \.bak
    | \.swp
    | \.log
    | \.sql
    | \.sqlite3
    | \.sqlite3-journal
    | ^setup\.py
    | ^__init__\.py
)
'''

[tool.isort]
line_length=79
include_trailing_comma = true
profile = "black"
py_version = 310
skip = [".gitignore", ".dockerignore", "__init__.py", "manage.py"]
skip_glob = [
    "docs/*",
    "**/migrations/*.py",
    "tests/",
    "*/migrations/",
    "venv/",
    "venv/*",
    "*/venv/",
    "env/",
    "static/*",
    "templates/*",
]

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "config.settings"